+++ BovineLabs.Anchor/App/NavigationStateSystem.cs
@@ -9,7 +9,6 @@ namespace BovineLabs.Anchor
     using Unity.Burst;
     using Unity.Collections;
     using Unity.Entities;
-    using UnityEngine;
 
     /// <summary>
     /// Synchronizes DOTS component states with the currently active Anchor navigation destination.
@@ -23,26 +22,18 @@ namespace BovineLabs.Anchor
         private NativeHashMap<FixedString32Bytes, ComponentType> statesMap;
 
         /// <inheritdoc/>
-        [BurstCompile]
         public void OnCreate(ref SystemState state)
         {
             this.statesMap = new NativeHashMap<FixedString32Bytes, ComponentType>(16, Allocator.Persistent);
-
-            var comps = new NativeHashSet<ComponentType>(0, Allocator.Temp);
 
             var e = UISystemTypes.Enumerator();
             while (e.MoveNext())
             {
                 var component = ComponentType.FromTypeIndex(TypeManager.GetTypeIndexFromStableTypeHash(e.Current.Value));
 
-                if (!comps.Add(component))
-                {
-                    continue;
-                }
-
                 if (!this.statesMap.TryAdd(e.Current.Name, component))
                 {
-                    Debug.LogError($"Key {e.Current.Name} has already been registered");
+                    BLGlobalLogger.LogErrorString($"Navigation state '{e.Current.Name}' has already been registered.");
                 }
             }
         }
@@ -64,12 +55,21 @@ namespace BovineLabs.Anchor
                 return;
             }
 
-            if (this.statesMap.TryGetValue(this.previous, out var previousComponent))
+            var hasPrevious = this.statesMap.TryGetValue(this.previous, out var previousComponent);
+            var hasCurrent = this.statesMap.TryGetValue(current, out var currentComponent);
+
+            if (hasPrevious && hasCurrent && previousComponent == currentComponent)
+            {
+                this.previous = current;
+                return;
+            }
+
+            if (hasPrevious)
             {
                 state.EntityManager.RemoveComponent(state.SystemHandle, previousComponent);
             }
 
-            if (this.statesMap.TryGetValue(current, out var currentComponent))
+            if (hasCurrent)
             {
                 state.EntityManager.AddComponent(state.SystemHandle, currentComponent);
             }